requirements.txt
# Core dependencies
numpy==1.26.4
pi-plates==10.5
websockets==10.4
scipy==1.15.1
aiofiles==24.1.0
typeguard==4.4.2

# RPI dependencies
RPLCD==1.3.1
smbus2==0.4.3
RPi.GPIO==0.7.1

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.5
